+++ packages/components/scripts/generateDSRComponents.ts
@@ -263,6 +263,11 @@ import { get${componentName}Css } from '${stylesBundleImportPath}';
           .replace(/this\.props\.(hasHeader|hasFooter|hasSubFooter)/g, '$1')
           .replace(/(?:hasHeader|hasFooter|hasSubFooter) =/g, 'const $&')
           .replace(/\n.*\/\/ eslint-disable-next-line @typescript-eslint\/member-ordering/, '');
+      } else if (tagName === 'p-radio-button-wrapper') {
+        newFileContent = newFileContent.replace(
+          /&& !(typeof otherChildren\[0] === 'object' && 'props' in otherChildren\[0]) && (otherChildren\[0]\?\.props\.checked)/g,
+          '&& !($1 && ($2 || otherChildren[0]?.props.defaultChecked))' // wrap in brackets because of negation
+        );
       } else if (tagName === 'p-tabs') {
         newFileContent = newFileContent
           .replace(/this\.tabsItemElements(\.map)/, `otherChildren$1`)