+++ packages/components/src/components.d.ts
@@ -1222,6 +1222,10 @@ export namespace Components {
          */
         "label"?: string;
         /**
+          * __Experimental__: Disables the radio button and shows a loading indicator.
+         */
+        "loading"?: boolean;
+        /**
           * The message styled depending on validation state.
          */
         "message"?: string;
@@ -3440,6 +3444,10 @@ declare namespace LocalJSX {
           * The label text.
          */
         "label"?: string;
+        /**
+          * __Experimental__: Disables the radio button and shows a loading indicator.
+         */
+        "loading"?: boolean;
         /**
           * The message styled depending on validation state.
          */